+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Ghaal'Shaarat
 - Perfect Weapon
 - Runes of War
+- Relentless Endurance
 ## API Update Notes
 - The icons used in `dialogUtils.selectDamageType` are now available under `constants.damageIcons`.