+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.5.22 Release:
+## Update Notes:
+- This is not a V14 update. An update for V14 is still a work in progress.
+## Bug Fixes
+- Wails from the Grave
+- Cleave Mastery
+- Graze Mastery
+- Topple Mastery
+
 # 1.5.21 Release:
 ## Update Notes:
 - This is not a V14 update. An update for V14 is still a work in progress.